Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/109-expected.txt

Not tested

FAIL  scheduler: javascript URL in iframe, src set via DOM assert_array_equals: expected property 2 to be "inline script #3" but got "iframe onload" (expected array ["inline script #1", "inline script #2", "inline script #3", "JS URL", "iframe onload"] got ["inline script #1", "inline script #2", "iframe onload", "inline script #3", "JS URL"])
